test: consolidate DFT tests into single file
- Merge 12 test files into tests/integrations/test_dft.py
- Organized by feature area with clear section headers
- All 83 tests pass (81 passed, 2 skipped)
- Update README.md to reflect new test structure
- Use AST parsing for syntax correctness

AST-based consolidation ensures:
- Proper import handling with aliases
- All helper functions preserved
- Emoji removal for syntax correctness
- Import deduplication and normalization
