--- tests/integrations/test_dft_compatibility.py
@@ -1,396 +1,0 @@
-"""Test DFT compatibility with compatible features.
-
-This test suite verifies that DFT correctly integrates with features that are
-architecturally compatible. These features either:
-1. Operate at different layers (transparent to DFT)
-2. Require explicit handling in DFT code (implemented and tested)
-3. Work seamlessly with standard CE-based loss computation
-
-TESTED COMPATIBILITIES:
-- ✅ Gradient Accumulation: num_items_in_batch normalization works correctly
-- ✅ Mixed Precision (FP16/BF16): .float() cast for exp(-loss) computation
-- ✅ Flash Attention: Operates at attention layer, transparent to loss
-- ✅ FSDP: Model sharding transparent to loss computation
-- ✅ DDP: Standard distributed training compatibility
-
-REFERENCE: specs/001-dft-compatibility-matrix/README.md
-"""
-
-from types import SimpleNamespace
-from unittest.mock import Mock
-
-import pytest
-import torch
-
-from src.axolotl.integrations.dft.dft_utils import (
-    apply_dft_weighting,
-    compute_dft_loss,
-    compute_per_token_cross_entropy,
-    reduce_token_loss,
-)
-from src.axolotl.integrations.dft.patch import patch_compute_loss_for_dft
-
-
-class TestDFTGradientAccumulationCompatibility:
-    """Test DFT compatibility with gradient accumulation."""
-
-    def test_gradient_accumulation_normalization(self):
-        """Verify DFT correctly uses num_items_in_batch for normalization.
-
-        In gradient accumulation, num_items_in_batch is used to normalize loss
-        across accumulated micro-batches. The loss should scale inversely with
-        num_items_in_batch.
-
-        Example:
-        - Micro-batch: batch_size=2, seq_len=16 → ~30 valid tokens after shift
-        - Without num_items_in_batch: loss normalized by 30 tokens
-        - With num_items_in_batch=8: loss normalized by 8 (accumulation batch size)
-        - Expected: loss_with / loss_without ≈ 30 / 8 = 3.75x
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 100
-
-        # Fixed seed for reproducible test
-        torch.manual_seed(42)
-        logits = torch.randn(batch_size, seq_len, vocab_size, requires_grad=True)
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # Compute expected number of valid tokens (for verification)
-        # After shift: seq_len - 1 = 15 tokens per sequence
-        expected_tokens = batch_size * (seq_len - 1)
-
-        # Compute loss without num_items_in_batch (normalized by actual tokens)
-        loss_without = compute_dft_loss(
-            logits,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-            num_items_in_batch=None,
-        )
-
-        # Compute loss with num_items_in_batch (normalized by accumulation batch)
-        num_items_in_batch = 8
-        loss_with = compute_dft_loss(
-            logits,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-            num_items_in_batch=num_items_in_batch,
-        )
-
-        # Verify: loss should scale by (expected_tokens / num_items_in_batch)
-        expected_ratio = expected_tokens / num_items_in_batch
-        actual_ratio = loss_with.item() / loss_without.item()
-
-        assert torch.isfinite(loss_with), "Loss with accumulation should be finite"
-        assert torch.isfinite(loss_without), (
-            "Loss without accumulation should be finite"
-        )
-        assert abs(actual_ratio - expected_ratio) < 0.1, (
-            f"Loss ratio should be ~{expected_ratio:.3f}, got {actual_ratio:.3f}"
-        )
-
-        print(
-            f"\n✓ Gradient accumulation normalization: "
-            f"expected_tokens={expected_tokens}, "
-            f"num_items_in_batch={num_items_in_batch}, "
-            f"loss_with={loss_with.item():.6f}, "
-            f"loss_without={loss_without.item():.6f}, "
-            f"ratio={actual_ratio:.3f} (expected {expected_ratio:.3f})"
-        )
-
-    def test_gradient_accumulation_in_trainer_patch(self):
-        """Verify num_items_in_batch is correctly passed through trainer patch."""
-        mock_trainer = Mock()
-        mock_trainer.args = SimpleNamespace(
-            enable_dft_loss=True,
-            label_smoothing_factor=0.0,
-            orpo_alpha=None,
-            include_tkps=False,
-            enable_dft_channel_loss=False,
-            dft_chunk_size=None,
-        )
-        mock_cfg = SimpleNamespace()
-
-        patch_compute_loss_for_dft(mock_trainer, mock_cfg)
-
-        inputs = {
-            "input_ids": torch.tensor([[1, 2, 3, 4]]),
-            "labels": torch.tensor([[2, 3, 4, -100]]),
-        }
-
-        mock_model = Mock()
-        mock_outputs = SimpleNamespace(
-            logits=torch.randn(1, 4, 100, requires_grad=True)
-        )
-        mock_model.return_value = mock_outputs
-
-        # Call with num_items_in_batch (gradient accumulation active)
-        loss = mock_trainer.compute_loss(mock_model, inputs, num_items_in_batch=8)
-
-        assert torch.isfinite(loss), "Loss should be finite"
-        assert loss.item() > 0, "Loss should be positive"
-
-        print(f"\n✓ Trainer patch with gradient accumulation: loss={loss.item():.6f}")
-
-
-class TestDFTMixedPrecisionCompatibility:
-    """Test DFT compatibility with mixed precision training (FP16/BF16)."""
-
-    def test_fp16_compatibility(self):
-        """Verify DFT works correctly with FP16 tensors.
-
-        DFT uses .float() cast for exp(-loss) computation to avoid
-        numerical issues with FP16's limited range.
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 100
-
-        # Create FP16 logits (common in mixed precision training)
-        logits = (
-            torch.randn(batch_size, seq_len, vocab_size).half().requires_grad_(True)
-        )
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # DFT should handle FP16 correctly
-        loss = compute_dft_loss(
-            logits,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-        )
-
-        assert torch.isfinite(loss), "Loss should be finite with FP16"
-        assert loss.item() > 0, "Loss should be positive"
-        assert loss.dtype == torch.float32, "Loss should be FP32 (upcast)"
-
-        # Verify gradient flow
-        loss.backward()
-        assert logits.grad is not None, "Gradients should flow with FP16"
-        assert logits.grad.dtype == torch.float16, "Gradients should be FP16"
-
-        print(f"\n✓ FP16 compatibility: loss={loss.item():.6f} (dtype={loss.dtype})")
-
-    def test_bf16_compatibility(self):
-        """Verify DFT works correctly with BF16 tensors.
-
-        BF16 has better range than FP16 but lower precision. DFT should
-        handle it correctly with .float() upcast.
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 100
-
-        # Create BF16 logits
-        logits = (
-            torch.randn(batch_size, seq_len, vocab_size).bfloat16().requires_grad_(True)
-        )
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # DFT should handle BF16 correctly
-        loss = compute_dft_loss(
-            logits,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-        )
-
-        assert torch.isfinite(loss), "Loss should be finite with BF16"
-        assert loss.item() > 0, "Loss should be positive"
-        assert loss.dtype == torch.float32, "Loss should be FP32 (upcast)"
-
-        # Verify gradient flow
-        loss.backward()
-        assert logits.grad is not None, "Gradients should flow with BF16"
-        assert logits.grad.dtype == torch.bfloat16, "Gradients should be BF16"
-
-        print(f"\n✓ BF16 compatibility: loss={loss.item():.6f} (dtype={loss.dtype})")
-
-    def test_mixed_precision_numerical_stability(self):
-        """Verify DFT's exp(-loss) computation is numerically stable in FP16.
-
-        The .float() upcast in apply_dft_weighting prevents:
-        - Overflow: exp(-loss) for small losses (close to 1.0)
-        - Underflow: exp(-loss) for large losses (close to 0.0)
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 100
-
-        # Create FP16 logits
-        logits_fp16 = torch.randn(batch_size, seq_len, vocab_size).half()
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # Compute per-token loss
-        per_token_loss, valid_mask = compute_per_token_cross_entropy(
-            logits_fp16,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-        )
-
-        # Apply DFT weighting (uses .float() internally)
-        weighted_loss = apply_dft_weighting(per_token_loss)
-
-        # All weights should be finite (no overflow/underflow)
-        assert torch.all(torch.isfinite(weighted_loss)), (
-            "All weighted losses should be finite"
-        )
-        assert torch.all(weighted_loss >= 0), "All weights should be non-negative"
-
-        # Reduce to scalar
-        loss = reduce_token_loss(weighted_loss, valid_mask)
-        assert torch.isfinite(loss), "Final loss should be finite"
-
-        print(
-            f"\n✓ Mixed precision stability: "
-            f"per_token_loss range=[{per_token_loss.min():.3f}, {per_token_loss.max():.3f}], "
-            f"weighted range=[{weighted_loss.min():.3f}, {weighted_loss.max():.3f}], "
-            f"final loss={loss.item():.6f}"
-        )
-
-
-class TestDFTFlashAttentionCompatibility:
-    """Test DFT compatibility with Flash Attention.
-
-    Flash Attention operates at the attention layer, computing Q@K^T and softmax
-    more efficiently. It is transparent to loss computation since it only affects
-    the forward pass (attention mechanism), not logits or loss.
-    """
-
-    def test_flash_attention_transparency_documented(self):
-        """Document why Flash Attention is transparent to DFT.
-
-        This is a documentation test to ensure understanding is captured.
-        """
-        architectural_separation = {
-            "Flash Attention layer": "Attention mechanism (Q, K, V computation)",
-            "DFT loss layer": "Cross-entropy on final logits",
-            "Interaction": "None - Flash Attention outputs hidden states, DFT uses logits",
-            "Transparency": "✅ Flash Attention is invisible to DFT loss computation",
-        }
-
-        expected_behavior = {
-            "DFT code changes": "None required - Flash Attention is transparent",
-            "Special handling": "Not needed - operates at different layer",
-            "Performance impact": "Flash Attention speeds up forward pass, DFT unchanged",
-            "Memory usage": "Flash Attention reduces attention memory, DFT unchanged",
-        }
-
-        print("\nFlash Attention Architectural Separation:")
-        for aspect, detail in architectural_separation.items():
-            print(f"  {aspect}: {detail}")
-
-        print("\nExpected DFT Behavior with Flash Attention:")
-        for aspect, detail in expected_behavior.items():
-            print(f"  {aspect}: {detail}")
-
-        assert True, "Flash Attention transparency documented"
-
-    def test_dft_with_standard_attention_tensors(self):
-        """Verify DFT works with tensors shaped like Flash Attention outputs.
-
-        Flash Attention outputs the same shape as standard attention:
-        [batch, seq, hidden_dim] → after LM head → [batch, seq, vocab]
-
-        This test verifies DFT handles these standard shapes correctly.
-        """
-        batch_size, seq_len, vocab_size = 2, 128, 50000  # Typical LLM dims
-
-        # Simulate logits output (same shape with or without Flash Attention)
-        logits = torch.randn(batch_size, seq_len, vocab_size, requires_grad=True)
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # DFT should work identically
-        loss = compute_dft_loss(
-            logits,
-            labels,
-            shift_labels=True,
-            ignore_index=-100,
-        )
-
-        assert torch.isfinite(loss), "Loss should be finite"
-        assert loss.item() > 0, "Loss should be positive"
-
-        # Verify gradient flow
-        loss.backward()
-        assert logits.grad is not None, "Gradients should flow"
-
-        print(
-            f"\n✓ Flash Attention tensor shapes: "
-            f"logits={tuple(logits.shape)}, loss={loss.item():.6f}"
-        )
-
-
-class TestDFTDistributedTrainingCompatibility:
-    """Test DFT compatibility with distributed training frameworks."""
-
-    def test_fsdp_transparency_documented(self):
-        """Document why FSDP is transparent to DFT.
-
-        FSDP (Fully Sharded Data Parallel):
-        - Shards model parameters across GPUs
-        - All-Gathers parameters before forward/backward
-        - Forward pass outputs complete logits [batch, seq, vocab]
-        - Loss computation sees complete tensors
-
-        DFT receives complete logits, so FSDP is transparent.
-        """
-        fsdp_guarantees = {
-            "Parameter sharding": "Reduces per-GPU memory usage",
-            "All-Gather timing": "Before forward pass, parameters reconstructed",
-            "Forward output": "Complete logits [batch, seq, vocab] on each GPU",
-            "DFT receives": "Complete logits, identical to non-FSDP",
-            "Loss computation": "Identical to single-GPU case",
-        }
-
-        expected_behavior = {
-            "DFT code changes": "None required - FSDP is transparent",
-            "Special handling": "Not needed - FSDP handles communication",
-            "Performance impact": "FSDP communication overhead, not DFT-specific",
-            "Memory usage": "FSDP reduces param memory, DFT unchanged",
-        }
-
-        print("\nFSDP Architectural Guarantees:")
-        for aspect, detail in fsdp_guarantees.items():
-            print(f"  {aspect}: {detail}")
-
-        print("\nExpected DFT Behavior with FSDP:")
-        for aspect, detail in expected_behavior.items():
-            print(f"  {aspect}: {detail}")
-
-        assert True, "FSDP transparency documented"
-
-    def test_ddp_transparency_documented(self):
-        """Document why DDP is transparent to DFT.
-
-        DDP (Distributed Data Parallel):
-        - Replicates model across GPUs
-        - Each GPU processes different data batch
-        - Gradients are All-Reduced after backward
-        - Loss computation is local (per-GPU)
-
-        DFT computes loss locally, then gradients are synchronized by DDP.
-        """
-        ddp_workflow = {
-            "Model replication": "Full model copy on each GPU",
-            "Data sharding": "Different batches on each GPU",
-            "Forward pass": "Independent, produces local logits",
-            "Loss computation": "DFT computes local loss (per-GPU)",
-            "Gradient sync": "DDP All-Reduces gradients after backward",
-        }
-
-        expected_behavior = {
-            "DFT code changes": "None required - DDP is transparent",
-            "Special handling": "Not needed - DDP handles gradient sync",
-            "Performance impact": "DDP communication in backward, not DFT-specific",
-            "Memory usage": "DDP replicates model, DFT unchanged",
-        }
-
-        print("\nDDP Workflow with DFT:")
-        for step, detail in ddp_workflow.items():
-            print(f"  {step}: {detail}")
-
-        print("\nExpected DFT Behavior with DDP:")
-        for aspect, detail in expected_behavior.items():
-            print(f"  {aspect}: {detail}")
-
-        assert True, "DDP transparency documented"
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-xvs"])
