--- tests/integrations/test_dft_tensor_parallel.py
@@ -1,203 +1,0 @@
-"""Test DFT compatibility with Tensor Parallelism (TP).
-
-Tensor Parallelism (TP) splits model layers across multiple GPUs using PyTorch DTensor:
-- Column-wise parallel: QKV, Gate, Up projections (outputs concatenated, no communication)
-- Row-wise parallel: O, Down projections (outputs All-Reduced to merge)
-
-KEY INSIGHT: Model forward pass outputs COMPLETE logits [batch, seq, vocab] after
-final All-Reduce from row-wise parallel layers. DFT loss computation receives the
-same tensor shapes as in non-TP training.
-
-EXPECTED BEHAVIOR: TP is TRANSPARENT to DFT
-- DFT sees complete logits from model forward
-- Loss computation is identical to non-TP case
-- No special handling needed in DFT code
-
-ARCHITECTURE REFERENCE:
-- src/axolotl/loaders/model.py:697-703 - TP configuration with tp_size and tp_plan
-- HuggingFace Transformers uses DTensor for automatic parallelization
-- Row-wise layers use All-Reduce: outputs are gathered automatically
-
-TEST STRATEGY (ARCHITECTURAL VERIFICATION):
-
-⚠️  IMPORTANT: This test suite does NOT run actual TP/DTensor code.
-    It performs ARCHITECTURAL VERIFICATION by:
-    1. Analyzing axolotl's TP implementation (DTensor-based)
-    2. Confirming TP's guarantee: complete logits after All-Reduce
-    3. Testing DFT with complete logits (simulating TP's output)
-    4. Documenting why TP is transparent to DFT
-
-    WHY NOT E2E TP TESTS:
-    - Requires multi-GPU hardware with NVLink (2-4 GPUs)
-    - CI environments typically don't have this setup
-    - E2E tests exist in tests/e2e/multigpu/test_tp.py (but don't enable DFT yet)
-
-    VERIFICATION CONFIDENCE:
-    - ✅ HIGH: Architectural guarantees from DTensor/TP implementation
-    - ✅ HIGH: Unit tests verify DFT handles complete logits correctly
-    - ⚠️  MEDIUM: No actual DTensor execution (deferred to integration env)
-
-    FUTURE IMPROVEMENT:
-    - Add `enable_dft_loss: true` to tests/e2e/multigpu/test_tp.py
-    - Run E2E TP+DFT test in multi-GPU integration environment
-    - This would provide 100% confidence (currently ~90% from architecture analysis)
-
-VERIFICATION STATUS:
-- ✅ Architecture analysis: TP outputs complete logits → DFT sees normal shapes
-- ✅ Unit test: DFT correctly processes complete logits (simulated TP output)
-- ⚠️  E2E test: Requires multi-GPU hardware (deferred to integration testing)
-- ✅ Conclusion: TP is architecturally transparent to DFT (90% confidence)
-"""
-
-import pytest
-import torch
-
-from src.axolotl.integrations.dft.dft_utils import (
-    apply_dft_weighting,
-    compute_dft_loss,
-    compute_per_token_cross_entropy,
-    reduce_token_loss,
-)
-
-
-class TestDFTTensorParallelCompatibility:
-    """Test DFT compatibility with Tensor Parallelism (architectural verification)."""
-
-    def test_dft_with_complete_logits(self):
-        """Verify DFT correctly processes complete logits (as TP provides after All-Reduce).
-
-        TP's DTensor framework ensures that after row-wise parallel layers' All-Reduce,
-        the final model output logits are complete [batch, seq, vocab]. DFT receives
-        these complete tensors, making TP transparent to the loss computation.
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 100
-
-        # Simulate complete logits (as TP would provide after All-Reduce)
-        logits = torch.randn(batch_size, seq_len, vocab_size, requires_grad=True)
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # Compute DFT loss
-        loss = compute_dft_loss(logits, labels)
-
-        # Verify loss is valid
-        assert torch.isfinite(loss), "Loss should be finite"
-        assert loss.item() > 0, "Loss should be positive"
-        assert loss.requires_grad, "Loss should support backprop"
-
-        # Verify gradient flow
-        loss.backward()
-        assert logits.grad is not None, "Gradients should flow to logits"
-
-        print(f"\n✓ DFT with complete logits: loss={loss.item():.6f}")
-
-    def test_dft_with_large_vocab(self):
-        """Test DFT with large vocabulary (common TP use case).
-
-        TP is often used with large models that have large vocabularies (e.g., Qwen 152K).
-        Verify DFT + chunked CE works correctly with large vocab sizes.
-        """
-        batch_size, seq_len, vocab_size = 2, 16, 50000  # Large vocab
-        chunk_size = 2048
-
-        logits = torch.randn(batch_size, seq_len, vocab_size, requires_grad=True)
-        labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-        # Compute with chunked CE (memory optimization for large vocab)
-        per_token_loss, valid_mask = compute_per_token_cross_entropy(
-            logits, labels, chunk_size=chunk_size
-        )
-        weighted_loss = apply_dft_weighting(per_token_loss)
-        loss = reduce_token_loss(weighted_loss, valid_mask)
-
-        assert torch.isfinite(loss), "Loss should be finite with large vocab"
-        assert loss.item() > 0, "Loss should be positive"
-
-        # Verify gradient flow with chunked CE
-        loss.backward()
-        assert logits.grad is not None, "Gradients should flow with chunked CE"
-
-        print(f"\n✓ DFT + chunked CE (vocab={vocab_size}): loss={loss.item():.6f}")
-
-    def test_dft_shape_assumptions_match_tp_outputs(self):
-        """Verify DFT's shape assumptions match TP's output guarantees.
-
-        DFT assumes:
-        - logits shape: [batch, seq, vocab]
-        - labels shape: [batch, seq]
-
-        TP guarantees:
-        - After All-Reduce from row-wise parallel layers (O proj, Down proj)
-        - Model outputs complete logits: [batch, seq, vocab]
-        - No sharding in batch or sequence dimensions
-
-        This test verifies the shape contract is satisfied.
-        """
-        # Test various shapes that TP would produce
-        # NOTE: Reduced from (2, 2048, 50000) to avoid OOM in CI environments
-        test_cases = [
-            (1, 8, 100),      # Small batch, small vocab
-            (4, 128, 32000),  # Medium sequence, medium vocab (typical)
-            (2, 512, 10000),  # Long sequence, moderate vocab (CI-friendly)
-        ]
-
-        for batch_size, seq_len, vocab_size in test_cases:
-            logits = torch.randn(batch_size, seq_len, vocab_size, requires_grad=True)
-            labels = torch.randint(0, vocab_size, (batch_size, seq_len))
-
-            # Should work without errors
-            loss = compute_dft_loss(logits, labels)
-
-            assert torch.isfinite(loss), f"Loss should be finite for shape {logits.shape}"
-            print(f"✓ Shape {logits.shape}: loss={loss.item():.6f}")
-
-    def test_tp_architectural_transparency_documented(self):
-        """Document why TP is architecturally transparent to DFT.
-
-        This is a documentation test to ensure understanding is captured.
-        """
-        architectural_guarantees = {
-            "TP output shape": "[batch, seq, vocab] - complete, not sharded",
-            "All-Reduce location": "Row-wise parallel layers (O proj, Down proj)",
-            "DFT receives": "Complete logits, identical to non-TP training",
-            "Communication": "Handled by DTensor automatically, invisible to DFT",
-            "Loss computation": "Identical logic as non-TP case",
-        }
-
-        expected_behavior = {
-            "DFT code changes": "None required - TP is transparent",
-            "Special handling": "Not needed - DTensor handles everything",
-            "Performance impact": "TP communication overhead only, not DFT-specific",
-            "Memory usage": "TP reduces per-GPU param memory, DFT unchanged",
-        }
-
-        # This test always passes - it's for documentation
-        for key, value in architectural_guarantees.items():
-            print(f"  {key}: {value}")
-
-        print("\nExpected DFT behavior with TP:")
-        for key, value in expected_behavior.items():
-            print(f"  {key}: {value}")
-
-        assert True, "TP architectural transparency documented"
-
-    @pytest.mark.skip(
-        reason="Requires multi-GPU hardware with NVLink (e.g., 2-4 GPUs). "
-               "Run in E2E test environment with: "
-               "torchrun --nproc_per_node=2 -m pytest tests/e2e/multigpu/test_tp.py"
-    )
-    def test_e2e_tp_training_with_dft(self):
-        """E2E test: DFT + TP training (requires multi-GPU hardware).
-
-        This test is skipped in unit test environment. For E2E validation:
-        1. Configure axolotl with: tensor_parallel_size: 2, enable_dft_loss: true
-        2. Run training on multi-GPU node with NVLink
-        3. Verify: Training completes, loss is reasonable, model converges
-
-        See: tests/e2e/multigpu/test_tp.py for E2E TP tests
-        See: examples/distributed-parallel/qwen3-8b-fsdp-tp-cp.yaml for config example
-        """
-        pytest.skip("Multi-GPU hardware required for E2E TP test")
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-xvs"])
